Deprecate old cancel unpaid orders command
